--- bridge-ui/src/hooks/transaction-args/cctp/useCCTPSrcDomain.ts
@@ -1,8 +1,0 @@
-import { useChainStore } from "@/stores";
-
-export const useCCTPSrcDomain = () => {
-  const fromChain = useChainStore.useFromChain();
-  return fromChain.cctpDomain;
-};
-
-export default useCCTPSrcDomain;
