--- bridge-ui/src/hooks/transaction-args/cctp/useCctpDestinationDomain.ts
@@ -1,8 +1,0 @@
-import { useChainStore } from "@/stores";
-
-export const useCCTPDestinationDomain = () => {
-  const toChain = useChainStore.useToChain();
-  return toChain.cctpDomain;
-};
-
-export default useCCTPDestinationDomain;
