+++ bridge-ui/src/hooks/transaction-args/cctp/useCctpUtilHooks.ts
@@ -1,8 +1,19 @@
-import useCCTPDestinationDomain from "./useCCTPDestinationDomain";
-import useCCTPSrcDomain from "./useCCTPSrcDomain";
+// Break pattern of 1 hook-1 file because TypeScript and CI were going nuts on filenames such as useCCTPDestinationDomain.ts
+
+import { useChainStore } from "@/stores";
 import { getCCTPFee } from "@/services/cctp";
 import { useQuery } from "@tanstack/react-query";
 import { CCTP_TRANSFER_MAX_FEE_FALLBACK, CCTP_TRANSFER_FEE_BUFFER } from "@/utils/cctp";
+
+export const useCCTPSrcDomain = () => {
+  const fromChain = useChainStore.useFromChain();
+  return fromChain.cctpDomain;
+};
+
+export const useCCTPDestinationDomain = () => {
+  const toChain = useChainStore.useToChain();
+  return toChain.cctpDomain;
+};
 
 export const useCCTPFee = (): bigint => {
   const srcDomain = useCCTPSrcDomain();
@@ -14,5 +25,3 @@ export const useCCTPFee = (): bigint => {
   if (!data) return CCTP_TRANSFER_MAX_FEE_FALLBACK;
   return BigInt(data.minimumFee) + CCTP_TRANSFER_FEE_BUFFER;
 };
-
-export default useCCTPFee;
