+++ bridge-ui/src/hooks/transaction-args/cctp/useDepositForBurnTxArgs.ts
@@ -3,8 +3,7 @@ import { useAccount } from "wagmi";
 import { encodeFunctionData, padHex, zeroHash } from "viem";
 import { useFormStore, useChainStore } from "@/stores";
 import { isCctp } from "@/utils/tokens";
-import useCCTPDestinationDomain from "./useCCTPDestinationDomain";
-import { useCCTPFee } from "./useCCTPFee";
+import { useCCTPFee, useCCTPDestinationDomain } from "./useCctpUtilHooks";
 import { CCTP_MIN_FINALITY_THRESHOLD } from "@/utils";
 
 type UseDepositForBurnTxArgs = {
